# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LINE WORKS 欠席・遅刻連絡アプリ
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ReportType = typeof ReportType[keyof typeof ReportType];


export const ReportType = {
  absence: 'absence',
  late: 'late',
  earlyLeave: 'earlyLeave',
  other: 'other',
} as const;

export type ReportStatus = typeof ReportStatus[keyof typeof ReportStatus];


export const ReportStatus = {
  sent: 'sent',
  failed: 'failed',
  pending: 'pending',
} as const;

export interface Report {
  id: number;
  type: ReportType;
  senderName: string;
  reason: string;
  /** 欠席・遅刻の日付 (YYYY-MM-DD) */
  date: string;
  /**
     * 遅刻の場合の到着予定時刻
     * @nullable
     */
  expectedTime?: string | null;
  /** @nullable */
  channelId?: string | null;
  status: ReportStatus;
  /** @nullable */
  lineWorksMessageId?: string | null;
  createdAt: string;
}

export type ReportInputType = typeof ReportInputType[keyof typeof ReportInputType];


export const ReportInputType = {
  absence: 'absence',
  late: 'late',
  earlyLeave: 'earlyLeave',
  other: 'other',
} as const;

export interface ReportInput {
  type: ReportInputType;
  /** @minLength 1 */
  senderName: string;
  /** @minLength 1 */
  reason: string;
  /** 欠席・遅刻の日付 (YYYY-MM-DD) */
  date: string;
  /** 遅刻の場合の到着予定時刻 */
  expectedTime?: string;
  /** 送信先チャンネルID（省略時はデフォルト） */
  channelId?: string;
}

export interface ReportSummary {
  total: number;
  absenceCount: number;
  lateCount: number;
  earlyLeaveCount: number;
  otherCount: number;
  recentReports: Report[];
}

export interface Channel {
  id: string;
  name: string;
  /** @nullable */
  type?: string | null;
}

export interface DigestResult {
  /** 送信した連絡件数 */
  sentCount: number;
  /** 対象日付 (YYYY-MM-DD) */
  date: string;
  /** @nullable */
  messageId?: string | null;
}

export interface ErrorResponse {
  error: string;
}

export type ListReportsParams = {
limit?: number;
offset?: number;
};