# lib/api-zod/src/generated/types/digestResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LINE WORKS 欠席・遅刻連絡アプリ
 * OpenAPI spec version: 0.1.0
 */

export interface DigestResult {
  /** 送信した連絡件数 */
  sentCount: number;
  /** 対象日付 (YYYY-MM-DD) */
  date: string;
  /** @nullable */
  messageId?: string | null;
}